[Google] Move the building methods into the BasicEvent object
